requirements.txt
pyahocorasick==1.1.4
torch==0.2.0.post3
